Move files to their appropriate locations. Still need to eventually extend the PiPiper::Driver class.
