+++ .travis.yml
@@ -1,4 +1,5 @@
 language: ruby
+cache: bundler
 rvm:
   - 1.9.3
   - 2.1.0
