--- lib/pi_piper/bcm2835/bcm2835.rb
@@ -1,4 +1,0 @@
-module PiPiper
-  class Bcm2835 < Driver
-  end
-end
