+++ lib/pi_piper/bcm2835/version.rb
@@ -1,5 +1,5 @@
 module PiPiper
-  class Bcm2835
-    VERSION = "0.1.0"
+  module Bcm2835
+    VERSION = '0.1.0'
   end
 end
