--- spec/pi_piper/bcm2835_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe PiPiper::Bcm2835 do
-  it 'has a version number' do
-    expect(PiPiper::Bcm2835::VERSION).not_to be nil
-  end
-end
